+++ org.eclipse.winery.model.tosca/src/main/java/org/eclipse/winery/model/tosca/HasId.java
@@ -1,0 +1,43 @@
+/*******************************************************************************
+ * Copyright (c) 2017 University of Stuttgart
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * and the Apache License 2.0 which both accompany this distribution,
+ * and are available at http://www.eclipse.org/legal/epl-v10.html
+ * and http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Contributors:
+ *    Oliver Kopp - initial code contribution
+ *******************************************************************************/
+package org.eclipse.winery.model.tosca;
+
+import javax.xml.bind.annotation.XmlAttribute;
+import javax.xml.bind.annotation.XmlID;
+import javax.xml.bind.annotation.XmlSchemaType;
+import javax.xml.bind.annotation.adapters.CollapsedStringAdapter;
+import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
+
+public abstract class HasId extends TExtensibleElements {
+
+	@XmlAttribute(name = "id", required = true)
+	@XmlJavaTypeAdapter(CollapsedStringAdapter.class)
+	@XmlID
+	@XmlSchemaType(name = "ID")
+	private String id;
+
+	public HasId() {
+	}
+
+	public HasId(String id) {
+		this.setId(id);
+	}
+
+	public String getId() {
+		return id;
+	}
+
+	public void setId(String value) {
+		this.id = value;
+	}
+
+}