+++ org.eclipse.winery.repository/src/test/java/org/eclipse/winery/repository/resources/servicetemplates/topologytemplates/TopologyTemplateResourceTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2012-2013 University of Stuttgart.
+ * Copyright (c) 2012-2017 University of Stuttgart.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * and the Apache License 2.0 which both accompany this distribution,
@@ -11,6 +11,10 @@
  *******************************************************************************/
 package org.eclipse.winery.repository.resources.servicetemplates.topologytemplates;
 
+import org.eclipse.winery.common.ids.definitions.ServiceTemplateId;
+import org.eclipse.winery.model.tosca.TTopologyTemplate;
+import org.eclipse.winery.repository.Utils;
+import org.eclipse.winery.repository.backend.Repository;
 import org.eclipse.winery.repository.resources.AbstractResourceTest;
 
 import org.junit.Test;
@@ -34,6 +38,35 @@ public class TopologyTemplateResourceTest  extends AbstractResourceTest {
 		this.setRevisionTo("3fe0df76e98d46ead68295920e5d1cf1354bdea1");
 		this.assertPut("servicetemplates/http%253A%252F%252Fwinery.opentosca.org%252Ftest%252Fservicetemplates%252Ffruits/baobab_serviceTemplate/topologytemplate/", "servicetemplates/baobab_topologytemplate_v2.json");
 		this.assertGet("servicetemplates/http%253A%252F%252Fwinery.opentosca.org%252Ftest%252Fservicetemplates%252Ffruits/baobab_serviceTemplate/topologytemplate/", "servicetemplates/baobab_topologytemplate_v2.json");
+	}
+
+	@Test
+	public void farmTopologyTemplateIsCorrectlyReturnAsJson() throws Exception {
+		this.setRevisionTo("2d35f0d3c15b384c53df10967164d97e4a7dd6f2");
+		this.assertGet("servicetemplates/http%253A%252F%252Fwinery.opentosca.org%252Ftest%252Fservicetemplates%252Ffruits/farm/topologytemplate/", "servicetemplates/farm_topologytemplate.json");
+	}
+
+	@Test
+	public void farmTopologyTemplateIsCorrectlyReturnedAsXml() throws Exception {
+		this.setRevisionTo("2d35f0d3c15b384c53df10967164d97e4a7dd6f2");
+		this.assertGet("servicetemplates/http%253A%252F%252Fwinery.opentosca.org%252Ftest%252Fservicetemplates%252Ffruits/farm/topologytemplate/", "servicetemplates/farm_topologytemplate.xml");
+	}
+
+	@Test
+	public void farmTopologyTemplateCanBeCreatedAsJson() throws Exception {
+		this.setRevisionTo("1e2054315f18e80c466c26e6918d6506ce53f7f7");
+
+		// Quick hack to ensure that the service template containing the tpology template exists
+		ServiceTemplateId id = new ServiceTemplateId("http://winery.opentosca.org/test/servicetemplates/fruits", "farm", false);
+		Repository.INSTANCE.flagAsExisting(id);
+
+		this.assertPut("servicetemplates/http%253A%252F%252Fwinery.opentosca.org%252Ftest%252Fservicetemplates%252Ffruits/farm/topologytemplate/", "servicetemplates/farm_topologytemplate.json");
+	}
+
+	@Test
+	public void farmTopologyTemplateJsonCanBeParsed() throws Exception {
+		final String jsonStr = AbstractResourceTest.readFromClasspath("servicetemplates/farm_topologytemplate.json");
+		final TTopologyTemplate topologyTemplate = Utils.mapper.readValue(jsonStr, TTopologyTemplate.class);
 	}
 
 }